Adding the Option to set Phy 3

+++ Laerdal.Dfu/Droid/DfuInstallation.cs
@@ -108,7 +108,7 @@ namespace Laerdal.Dfu
             // Phy
             if (Phy.HasValue)
             {
-                Initiator = Initiator.SetPhy(Phy.Value);
+                Initiator = Initiator.SetPhy((int) Phy.Value);
             }
             
             // NumberOfRetries
